- all test passed

+++ src/main/java/com/nasnav/dao/RoleEmployeeUserRepository.java
@@ -14,7 +14,7 @@ public interface RoleEmployeeUserRepository extends JpaRepository<RoleEmployeeUs
     @Query("select roleEmp from RoleEmployeeUser roleEmp where roleEmp.employeeUserId = :employeeUserId")
     List<RoleEmployeeUser> findRoleEmployeeUsersById(@Param("employeeUserId") Integer employeeUserId);
 
-    boolean existsByEmployeeUserIdAndRoleId(Integer employeeUserId, Integer roleId);
+    boolean existsByEmployeeUserIdAndRoleId(Long employeeUserId, Integer roleId);
 
     @Transactional
     void deleteByEmployeeUserId(Long employeeUserId);
